add raed file

+++ edu/file_with_go/refactor_on_pos.go
@@ -4,13 +4,13 @@ import (
 	"os"
 )
 
-func ModifyFile(filename string, pos int64, val string) error {
+func ModifyFile(filename string, pos int, val string) error {
 	file, err := os.OpenFile(filename, os.O_RDWR, 0644)
 	if err != nil {
 		return err
 	}
 	defer file.Close()
-	if _, err := file.Seek(pos, 0); err != nil {
+	if _, err := file.Seek(int64(pos), 0); err != nil {
 		return err
 	}
 	if _, err := file.Write([]byte(val)); err != nil {
@@ -18,3 +18,8 @@ func ModifyFile(filename string, pos int64, val string) error {
 	}
 	return nil
 }
+
+// В этой функции мы открываем файл с помощью os.OpenFile с
+// флагом os.O_RDWR, что позволяет нам читать и записывать данные.
+// Затем мы перемещаемся к указанной позиции в файле с использованием file.Seek,
+// и, наконец, записываем значение val в эту позицию с помощью file.Write.
